refactor: move custom types to separate file
